+++ test/test-cases/aliases/expected.css
@@ -1,10 +1,13 @@
-._aliases_borders__bar {
-  color: red;
+._aliases_borders__dashed {
+  border: 4px dashed;
 }
+
+
 ._aliases_source__foo {
   background: aquamarine;
   border-color: red;
 }
+
 @media (max-width: 599px) {
   ._aliases_source__foo {
     background: white;
