+++ test/test-cases/aliases/source.css
@@ -1,4 +1,4 @@
-@let borders: "./borders.css", breakpoints: "./breakpoints";
+@let borders: "./borders.css", breakpoints: "./breakpoints.css";
 @import small from breakpoints;
 @import secondary, primary as blue from "./colors.css";
